Synchronize kernel.mk with buildroot, fix .configured target (no copy of .config file)
git-svn-id: http://svn.freetz.org/trunk@532 f5190166-0702-4917-9039-51ec32eddaf5

+++ make/linux/kernel.mk
@@ -62,22 +62,30 @@ endif
 $(KERNEL_DIR)/.configured: $(KERNEL_DIR)/.unpacked $(KERNEL_CONFIG_FILE)
 	cp $(KERNEL_CONFIG_FILE) $(KERNEL_BUILD_DIR)/kernel/linux-2.6.13.1/.config
 	export PATH=$(KERNEL_MAKE_PATH):$(PATH); \
-	$(MAKE1) -C $(KERNEL_BUILD_DIR)/kernel/linux-2.6.13.1 \
+	$(MAKE) -C $(KERNEL_BUILD_DIR)/kernel/linux-2.6.13.1 \
 		CROSS_COMPILE="$(KERNEL_CROSS)" \
 		KERNEL_MAKE_PATH="$(KERNEL_MAKE_PATH):$(PATH)" \
 		ARCH="$(KERNEL_ARCH)" \
 		KERNEL_LAYOUT="$(KERNEL_BOARD_REF)" \
-		oldconfig prepare
-	cp $(KERNEL_BUILD_DIR)/kernel/linux-2.6.13.1/.config $(KERNEL_CONFIG_FILE)
+		oldconfig 
+	touch $@
+	
+$(KERNEL_DIR)/.depend_done:  $(KERNEL_DIR)/.configured
+		$(MAKE) -C $(KERNEL_BUILD_DIR)/kernel/linux-2.6.13.1 \
+		CROSS_COMPILE="$(KERNEL_CROSS)" \
+		KERNEL_MAKE_PATH="$(KERNEL_MAKE_PATH):$(PATH)" \
+		ARCH="$(KERNEL_ARCH)" \
+		KERNEL_LAYOUT="$(KERNEL_BOARD_REF)" \
+		prepare
 	touch $@
 
-$(KERNEL_BUILD_DIR)/$(KERNEL_LZMA_LIB): $(KERNEL_DIR)/.configured $(TOOLS_DIR)/lzma
+$(KERNEL_BUILD_DIR)/$(KERNEL_LZMA_LIB): $(KERNEL_DIR)/.depend_done $(TOOLS_DIR)/lzma
 	export PATH=$(KERNEL_MAKE_PATH):$(PATH); \
 	$(MAKE) -C $(KERNEL_BUILD_DIR)/lzma lzma_decode.a CROSS_COMPILE=$(KERNEL_CROSS) \
 		USE_CFLAGS="$(KERNEL_LZMA_CFLAGS)"
 	cp $(KERNEL_BUILD_DIR)/lzma/lzma_decode.a $@
 
-$(KERNEL_BUILD_DIR)/$(KERNEL_IMAGE): $(KERNEL_DIR)/.configured $(KERNEL_BUILD_DIR)/$(KERNEL_LZMA_LIB) $(TOOLS_DIR)/lzma2eva
+$(KERNEL_BUILD_DIR)/$(KERNEL_IMAGE): $(KERNEL_DIR)/.depend_done $(KERNEL_BUILD_DIR)/$(KERNEL_LZMA_LIB) $(TOOLS_DIR)/lzma2eva
 	export PATH=$(KERNEL_MAKE_PATH):$(PATH); \
 	$(MAKE) -C $(KERNEL_BUILD_DIR)/kernel/linux-2.6.13.1 \
 		CROSS_COMPILE="$(KERNEL_CROSS)" \
@@ -125,15 +133,16 @@ kernel-configured: $(KERNEL_DIR)/.configured
 
 kernel-modules: $(KERNEL_DIR)/.modules
 
-kernel-menuconfig: $(KERNEL_DIR)/.unpacked $(KERNEL_CONFIG_FILE)
-	cp $(KERNEL_CONFIG_FILE) $(KERNEL_BUILD_DIR)/kernel/linux-2.6.13.1/.config
-	$(MAKE) -C $(KERNEL_BUILD_DIR)/kernel/linux-2.6.13.1 \
+kernel-menuconfig: $(KERNEL_DIR)/.unpacked
+	[ -f $(KERNEL_BUILD_DIR)/kernel/linux-2.6.13.1/.config ] || cp $(KERNEL_CONFIG_FILE) $(KERNEL_BUILD_DIR)/kernel/linux-2.6.13.1/.config
+		$(MAKE) -C $(KERNEL_BUILD_DIR)/kernel/linux-2.6.13.1 \
 		CROSS_COMPILE="$(KERNEL_CROSS)" \
 		KERNEL_MAKE_PATH="$(KERNEL_MAKE_PATH):$(PATH)" \
 		ARCH="$(KERNEL_ARCH)" \
 		KERNEL_LAYOUT="$(KERNEL_BOARD_REF)" \
 		menuconfig
-	cp $(KERNEL_BUILD_DIR)/kernel/linux-2.6.13.1/.config $(KERNEL_CONFIG_FILE)
+	-cp -f $(KERNEL_BUILD_DIR)/kernel/linux-2.6.13.1/.config $(KERNEL_CONFIG_FILE) && \
+	touch $(KERNEL_DIR)/.configured
 
 kernel-oldconfig: kernel-configured
